third_party/WebKit/LayoutTests/fast/dom/shadow/link-in-shadow-tree-expected.txt

CONSOLE WARNING: HTML element <link> is ignored in shadow tree.
PASS computedBackgroundColor(spanInShadow) is "rgba(0, 0, 0, 0)"
PASS computedBackgroundColor(spanInHostChildren) is "rgba(0, 0, 0, 0)"
PASS computedBackgroundColor(spanOutsideOfShadow) is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
tehepero
tehepero
